+++ _test.py
@@ -7,20 +7,93 @@ from cli_user import *
 
 
 # =====================================================================================================================
+if "Windows" in platform.system():
+    CMD_PING_1 = "ping -n 1 localhost"
+    CMD_PING_2 = "ping -n 2 localhost"
+else:
+    CMD_PING_1 = "ping -c 1 localhost"
+    CMD_PING_2 = "ping -c 2 localhost"
+
+
+# =====================================================================================================================
 class Test:
     def test__ok(self):
         victim = CliUser()
 
-        if "Windows" in platform.system():
-            cmd_line = "ping -n 1 localhost"
-        else:
-            cmd_line = "ping -c 1 localhost"
-
-        assert victim.send(cmd_line, timeout=2)
-        assert victim.last_cmd == cmd_line
+        assert victim.send(CMD_PING_1, timeout=2)
+        assert victim.last_cmd == CMD_PING_1
         assert victim.last_finished is True
 
-        assert victim._last_exx_timeout is None
+        assert victim.last_exx_timeout is None
+        assert bool(victim.last_stdout) is True
+        assert bool(victim.last_stderr) is False
+        assert victim.last_retcode == 0
+        assert victim.last_finished_success
+        assert victim.counter == 1
+        assert victim.counter_in_list == 0
+
+    def test__count(self):
+        victim = CliUser()
+        assert victim.counter == 0
+        assert victim.counter_in_list == 0
+
+        assert victim.send(CMD_PING_1, timeout=1)
+        assert victim.counter == 1
+        assert victim.counter_in_list == 0
+
+        assert victim.send(CMD_PING_1, timeout=1)
+        assert victim.counter == 2
+        assert victim.counter_in_list == 0
+
+        assert victim.send([CMD_PING_1, CMD_PING_1], timeout=1)
+        assert victim.counter == 4
+        assert victim.counter_in_list == 2
+
+        assert victim.send(CMD_PING_1, timeout=1)
+        assert victim.counter == 5
+        assert victim.counter_in_list == 0
+
+        assert victim.send([CMD_PING_1, CMD_PING_1, CMD_PING_1], timeout=1)
+        assert victim.counter == 8
+        assert victim.counter_in_list == 3
+
+        assert not victim.send([CMD_PING_2, CMD_PING_2, CMD_PING_2], timeout=1)
+        assert victim.counter == 9
+        assert victim.counter_in_list == 1
+
+    def test__list(self):
+        victim = CliUser()
+
+        assert not victim.send([CMD_PING_1, CMD_PING_2], timeout=1)
+        assert victim.counter_in_list == 2
+
+        assert not victim.send([CMD_PING_1, CMD_PING_2, CMD_PING_2], timeout=2)
+        assert victim.counter_in_list == 3
+
+        assert victim.send([CMD_PING_1, CMD_PING_2], timeout=2)
+        assert victim.counter_in_list == 2
+        assert victim.last_cmd == CMD_PING_2
+        assert victim.last_finished is True
+
+        assert victim.last_exx_timeout is None
+        assert bool(victim.last_stdout) is True
+        assert bool(victim.last_stderr) is False
+        assert victim.last_retcode == 0
+        assert victim.last_finished_success
+
+    def test__list__till_first_true(self):
+        victim = CliUser()
+
+        assert not victim.send([CMD_PING_1, CMD_PING_2], timeout=1)
+        assert victim.counter_in_list == 2
+
+        assert victim.send([CMD_PING_1, CMD_PING_2], timeout=1, till_first_true=True)
+        assert victim.counter_in_list == 1
+
+        assert victim.last_cmd == CMD_PING_1
+        assert victim.last_finished is True
+
+        assert victim.last_exx_timeout is None
         assert bool(victim.last_stdout) is True
         assert bool(victim.last_stderr) is False
         assert victim.last_retcode == 0
@@ -34,7 +107,7 @@ class Test:
         assert victim.last_cmd == cmd_line
         assert victim.last_finished is True
 
-        assert isinstance(victim._last_exx_timeout, Exx_CliTimeout)
+        assert isinstance(victim.last_exx_timeout, Exx_CliTimeout)
         assert bool(victim.last_stdout) is False
         assert bool(victim.last_stderr) is False
         assert victim.last_retcode is None
@@ -48,7 +121,7 @@ class Test:
         assert victim.last_cmd == cmd_line
         assert victim.last_finished is True
 
-        assert victim._last_exx_timeout is None
+        assert victim.last_exx_timeout is None
         # assert bool(victim.last_stdout) is False
         assert bool(victim.last_stderr) is True
         assert victim.last_retcode not in [0, None]
@@ -65,13 +138,8 @@ class Test:
             assert isinstance(exx, Exx_CliNotAvailable)
 
         # two cmd ------------------------------------------------
-        if "Windows" in platform.system():
-            cmd_ping_1 = "ping -n 1 localhost"
-        else:
-            cmd_ping_1 = "ping -c 1 localhost"
-
         class CliUserForAvailable(CliUser):
-            CMDS_REQUIRED = {cmd_ping_1: None, }
+            CMDS_REQUIRED = {CMD_PING_1: None, }
 
         victim = CliUserForAvailable()
         assert victim.cli_check_available()