+++ cli_user/main.py
@@ -1,6 +1,7 @@
 import subprocess
 import datetime
 import platform
+import time
 
 from typing import *
 
@@ -90,22 +91,27 @@ class CliUser:
     :exception Exx_CliRetcode:
     :exception Exx_CliStderr:
     """
+    # SETTINGS ------------------------------------
     TIMEOUT: Optional[float] = 2
     RAISE: Optional[bool] = None
 
     CMDS_REQUIRED: Optional[Dict[str, Optional[str]]] = None
 
+    _buffer_indent: str = "\t|"
+
+    # VALUES --------------------------------------
     _last_sp: Optional[subprocess.Popen] = None
-    _last_exx_timeout: Optional[Exx_CliTimeout] = None
+
+    counter: int = 0
+    counter_in_list: int = 0
 
     last_duration: float = 0
     last_cmd: str = ""
     last_stdout: str = ""         # USE ONLY "" AS DEFAULT!!!
     last_stderr: str = ""         # USE ONLY "" AS DEFAULT!!!
+    last_exx_timeout: Optional[Exx_CliTimeout] = None
     last_retcode: Optional[int] = None
     last_finished: Optional[bool] = None
-
-    _buffer_indent: str = "\t|"
 
     # init ------------------------------------------------------------------------------------------------------------
     def __init__(self, *args, **kwargs):
@@ -118,35 +124,76 @@ class CliUser:
 
     def _reinit_values(self) -> None:
         self._last_sp = None
-        self._last_exx_timeout = None
 
         self.last_duration = 0
         self.last_cmd = ""
         self.last_stdout = ""
         self.last_stderr = ""
+        self.last_exx_timeout = None
         self.last_retcode = None
         self.last_finished = None
 
+    @property
+    def last_finished_success(self) -> Optional[bool]:
+        """check if process last_finished correct
+
+        :return: True if last_finished with no errors
+        """
+        return all([self.last_finished, self.last_retcode == 0, not self.last_stderr, not self.last_exx_timeout])
+
     # SEND ------------------------------------------------------------------------------------------------------------
-    def send(self, cmd: Union[str, List[str]], timeout: Optional[float] = None, _raise: Optional[bool] = None) -> Union[bool, NoReturn]:
+    def send(
+            self,
+            cmd: Union[str, List[str]],
+            timeout: Optional[float] = None,
+            till_first_true: Optional[bool] = None,
+            _raise: Optional[bool] = None,
+            _use_counter_list: Optional[bool] = None
+    ) -> Union[bool, NoReturn]:
         """execute CLI command in terminal
 
         :param cmd: commands for execution
-        :param timeout: use special timeout step_result_enum, instead of default
+        :param timeout: use special timeout step_result_enum, instead of default, for cms_list will used as cumulated!
+        :param till_first_true: useful for detection or just multiPlatform usage
+        :param _raise: if till_first_true=True it will not work (return always bool in this case)!!!
+        :param _use_counter_list: DONT USE! internal flag for counter_in_list
         """
         # CMDS LIST ---------------------------------------------------------------------------------------------------
-        if isinstance(cmd, (list, )):
+        if isinstance(cmd, (list, tuple, )):
+            self.counter_in_list = 0
+            time_start = time.time()
+            result = True
+
+            _raise_list = _raise
+            if till_first_true:
+                _raise_list = False
+
             for cmd_item in cmd:
-                if not self.send(cmd=cmd_item, timeout=timeout, _raise=_raise):
-                    msg = f"[ERROR] {cmd_item=} in full sequence {cmd=}"
-                    print(msg)
-                    return False
+                self.counter_in_list += 1
+                time_passed = time.time() - time_start
+                timeout = timeout - time_passed
+
+                result = self.send(cmd=cmd_item, timeout=timeout, _raise=_raise_list, _use_counter_list=True)
+
+                if till_first_true:
+                    if result:
+                        return True
+                else:
+                    if not result:
+                        msg = f"[ERROR] {cmd_item=} in full sequence {cmd=}"
+                        print(msg)
+                        return False
+
+            return result
 
         # params reinit -----------------------------------------------------------------------------------------------
         self._reinit_values()
+        if not _use_counter_list:
+            self.counter_in_list = 0
 
         # params apply ------------------------------------------------------------------------------------------------
         self.last_cmd = cmd
+        self.counter += 1
 
         if timeout is None:
             timeout = self.TIMEOUT
@@ -154,6 +201,14 @@ class CliUser:
             _raise = self.RAISE
 
         # work --------------------------------------------------------------------------------------------------------
+        if timeout < 0:
+            msg = f"{timeout=} is sub zero"
+            self.last_exx_timeout = Exx_CliTimeout(msg)
+            if _raise:
+                raise Exx_CliTimeout(msg)
+            else:
+                return False
+
         # todo: check for linux! encoding is not necessary!
         time_start = datetime.datetime.now()
         self._last_sp = subprocess.Popen(args=cmd, text=True, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)    #, encoding="cp866"
@@ -164,15 +219,15 @@ class CliUser:
             self.last_stderr = self._last_sp.stderr.read()
             self.last_retcode = self._last_sp.returncode
         except subprocess.TimeoutExpired as exx:
-            self._last_exx_timeout = Exx_CliTimeout(repr(exx))
+            self.last_exx_timeout = Exx_CliTimeout(repr(exx))
             self.last_duration = (datetime.datetime.now() - time_start).total_seconds()
 
         self.last_finished = True
         self.print_state()
 
         if _raise:
-            if self._last_exx_timeout:
-                raise self._last_exx_timeout
+            if self.last_exx_timeout:
+                raise self.last_exx_timeout
             if self.last_retcode != 0:
                 raise Exx_CliRetcode(self.last_retcode)
             if self.last_stderr:
@@ -181,18 +236,14 @@ class CliUser:
         return self.last_finished_success
 
     # AUXILIARY -------------------------------------------------------------------------------------------------------
-    @property
-    def last_finished_success(self) -> Optional[bool]:
-        """check if process last_finished correct
-
-        :return: True if last_finished with no errors
-        """
-        return all([self.last_finished, self.last_retcode == 0, not self.last_stderr, not self._last_exx_timeout])
-
     def print_state(self) -> None:
         print(f"="*50)
         if not self.last_finished_success:
             print(f"[{'#'*21}ERROR{'#'*21}]")
+
+        print(f"{self.counter=}")
+        print(f"{self.counter_in_list=}")
+
         print(f"{self.last_cmd=}")
         print(f"{self.last_duration=}")
         print(f"{self.last_finished=}")
@@ -209,7 +260,7 @@ class CliUser:
             for line in self.last_stderr.split("\n"):
                 print(f"{self._buffer_indent}{line!r}")
         print(f"-" * 50)
-        print(f"{self._last_exx_timeout=}")
+        print(f"{self.last_exx_timeout=}")
         print(f"="*50)
 
     def cli_check_available(self) -> bool: